voice-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
coqui-tts==0.22.1
numpy==1.26.3
pydantic==2.6.0
python-multipart==0.0.9
transformers==4.42.4
torch==2.2.0
torchaudio==2.2.0
